Add runtime option IGNORE_MIN_REQUIREMENTS to ignore system requirement checks (#11)
This change exposes an IGNORE_MIN_REQUIREMENTS Docker environment variable
that allows the user to disable the minimum system requirements check. Per
[this MemSQL article](https://help.memsql.com/hc/en-us/articles/115001215583-My-hosts-have-less-than-minimum-MemSQL-system-requirements-How-can-I-make-MemSQL-run-on-those-hosts-) the article below this is not a recommended practice
for development or production. I've added this caveat in the README so users
are informed before choosing to ignore this check.

Implements https://github.com/memsql/memsql-docker-quickstart/issues/10